Use while let for run method
With a for loop, we have to clone queue, meaning that when we
crawl a url, the for loop doesn't see the new queue, just the old one.
With while let, the program will continue to get every URL from queue until it is empty, meaning we
can now scrape an entire site ourselves.